+++ include/copal/render.h
@@ -6,6 +6,7 @@
 #include <copal/types.h>
 #include <copal/theme.h>
 #include <copal/font.h>
+#include <copal/image.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -28,6 +29,10 @@ CL_API void cl_paint_stroke_round_rect(cl_paint_context_t *ctx, cl_rect_t r,
 CL_API void cl_paint_draw_text(cl_paint_context_t *ctx, cl_font_t *font,
                                const char *utf8, cl_point_t pos,
                                cl_color_t color);
+
+/** cl_paint_draw_image() - blend an image into dst, scaling as needed. */
+CL_API void cl_paint_draw_image(cl_paint_context_t *ctx, cl_image_t *img,
+                                cl_rect_t dst);
 
 /*
  * Clip region stack. cl_paint_push_clip() intersects r with the current clip